:truck: Moves animated image to atoms
