+++ pages/read/index.tsx
@@ -4,7 +4,7 @@ import makeStyles from "@material-ui/core/styles/makeStyles";
 import Typography from "@material-ui/core/Typography";
 import Button from "@material-ui/core/Button";
 import Grid from "@material-ui/core/Grid";
-import {ConstructionAnimation} from "../../src/component/AnimatedImage";
+import {ConstructionAnimation} from "../../src/component/atoms/AnimatedImage";
 
 const useStyles = makeStyles((theme) => ({
     root: {
